LayoutTests/fast/html/find-selects-draggable-element-expected.txt

PASS window.getSelection().toString() == 'Something' is true
Something
